Don't decorate user tasks, channels and RTIs.
Until now user task names are internally prefixed with '~' and
system task names are prefixed with '!' so appropriate scheduler
is used at remote peers. However, this can be confusing to users;
so, leave user task names, channel names and RTI names as they are
and only decorate system tasks (with '~').