Fixed obtaining server task reference.
Server task reference can be deserialized only if the peer is
already known (in networked pycos) - otherwise, location is set to
None. Now instead of sending task reference, location reference is
sent in multiprocessing queue and node then first establishes that
location as peer and then 'locate's server task.

The servers also send their task reference to node (for redundancy)
with 'server_task' message as soon as possible, and also when they
are finished with 'server_done' message. To prevent conflicts with
redundant methods of getting these messages, they are associated
with order IDs.