Added 'zombie_period' parameter to Computation.
Parameter 'zombie_period' to Computation indicates upper limit to
how long (in seconds) a node can lose communication to determine
it is a zombie (i.e., considered not usable and abandoned). This
setting is valid only for "prviate" schedulers; shared schedulers
should be started with 'zombie_period' option instead.

If a node is restarted, previous incarnation is discarded and new
one is used. However, fault tolerance is (still) not supported by
the scheduler (e.g., to restart tasks on the new node).