Changed semantics of 'tcp_port'.
Now the default value of 'tcp_port' is None instead of 9705. With
this default value, netpycos tries to use 9705 if possible. If that
port is already in use (likely another pycos program, such as a
server), a random port is chosen (as if 'tcp_port=0' is given).
With this, two different pycos programs (such as server and client)
can be run on same computer.

Updated version to 4.7.0 (due to change in network ports used).