Simplified Singleton metaclass.
Changed Singleton metaclass so class instances are referred to
using Singleton itself instead of given class 'cls', as seem to
be general accepted practice. However, using 'cls' is confusing
(at least not straigt forward and takes a bit of effort to have
an idea of why it works). In any case, instances are never stored
in 'cls' even if that is implied - it is just that object resolution
eventually finds it back in Singleton class (unlike in the case
when instances are stored with a variable and not a dictionary,
as done in earlier versions of pycos, in which case the variable
is stored in 'cls' when it is assigned).