Fixed dispycosnode for Windows.
Apparently with Python 2 '_dispycos_server_proc*' functions should
be in global scope (so they should not be removed in 'spawn'
function). This used to work until couple of releases ago (when
dispycosnode was tested with Windows). I don't know what broke it.
For now same changes are applied to Python 3 implementation as well.