Make sure remote references are valid.
Tasks and channels are associated with unique IDs, obtained with
builtin 'id' function. This remains unique for the life of object,
but if the object goes away (e.g., task is done) and a new object
is created, it is likely new object may get previously used IDs.
So there is a chance that if an object is used remotely, when that
reference is received back at the source, that object's ID may be
meant for different object. Now additional attribute is used in
remote references so such references can be validted at the source.