Moved node implementation from 'main' to function.
Implementation of node functionality has been moved from '__main__'
to a function so variables become locals (in '__main__', all
variables are prefixed with '_dispycos_' that are then removed
in child process so they are not visible to client computations).

Now server process IDs are sent to node so even if client programs
alter / remove files that contain IDs, node can terminate those
processes if necessary. When terminating processes, if psutil
module is available, the process IDs are used to obtain process
name to make sure the process is dispyconode (and not any other
process).