If node is busy, remove from peers.
If node is busy (i.e,. reserving fails), remove node from peers
so when node broadcasts it (when it is released by current client),
it can be reinitialized.

This approach works only for clients on local network; if client is
in remote network, clients have to periodically call 'peer' method
to use node, which is cumbersome / inefficient. However, this should
be rare case, so it is better to leave it to clients to implement.