Support Python 3.7
Python 3.7 changes 'raise StopIteration' semantics, which breaks
pycos framework. However, due to the way pycos task scheduler
handles tasks, it is a simple fix that can be done with 'sed'.
Currently this is done outside so not committing changes in git;
instead, the files are updated in releases. It would've been nice
if this can be done with setuptools (apparently overriding 'install'
in setup should be possible, but I can't get it working).