Check if process is_alive() before terminating process.
Don't rely on server.status to know if process is done, as there
can be a delay after process is done and server.status is set to
None. Otherwise, it may produce spurious error message (although
no harm done).

Fixed typo in previous commit ('dispycos_path' is not defined in
spawn process).