Changed zombie_period semantics.
If zombie_period is > 0, it is assumed servers and clients can be
closed, which is useful when nodes are shared. If zombie_period = 0
(which is default now), servers and clients are not closed unless
clients set 'abandon_zombie_nodes=True'. Servers are closed if no
tasks are running for zombie_period and clients are closed if no
servers are used for 2*zombie_period. Now clients are closed if
scheduler reserves but no server is used (e.g., if scheduler is
not reachable after reserving a node).