Fixed processing "quit" command.
Main thread's sighandler should not rely on node task's alive
check to send "quit" command to that task or to raise SIGINT -
it is likely that node task is about to quit (which would've
caused sighandler to fire) but before "quit" message is sent, the
task is done. Instead, remove pid file in node task and check for
it in sighandler.