Replaced MonitorException with MonitorStatus.
MonitorException was originally intended to be thrown as an
exception in monitor task. However, soon after it was changed to
be sent as message. For backward compatability, MonitorException
was kept same, but it's args are error prone in using them.

Now this has been replaced with MonitorStatus class which can be
used for task status message as well as informing exceptions.

Since this changes user API, pycos version has been bumped to
4.10.0.