Avoid serializing result.
Instead of checking if task result can be serialized before
sending it to scheduler, use 'deliver' instead of 'send' and
check its result. If 'deliver' fails, it may be likely due to
serialization error (although it can be due to other reasons as
well). If 'deliver' fails, check for serialization and send
appropriate message to scheduler.