Added dispycos_mem_server.py example.
Added example to show how dispycos can be used to load data in
each server's memory and query computations on this data
repeatedly. In essence, this allows client to view each CPU of
cluster to be viewed as separate computer (with each dispycos
server that runs on a CPU having its own network address, memory).
In this example same computation is used for simplification, but
if necessary each server can run separate computation or one server
can run multiple computations.