Set "value" of failed task to MonitorStatus.
Currently if a task fails, its "value" is None, which can be
misleading if None is valid "exit" value of a task and not very
useful even otherwise. Now "value" is set to MonitorStatus instance
with exception information so caller / monitor can use the result
to check if a task failed (by checking if it an instance of
MonitorStatus). This is done for both local tasks and remote tasks
that are received directly from dispycos or RPS.