Made scheduler tiny bit efficient with 3.7+.
With Python 3.7+ 'exit' value from tasks are always tuples of
size 1. This is different from earlier versions so it required
a test of Python version to process the results.
When installing with 'pip' of Python 3.7+, there is no need to
test Python version, so this setup.py translates the test to
'True'.